Add evidence pipeline, dashboard, and CAD cache

+++ firmware/esp32/src/drivers/encoder.cpp
@@ -67,7 +67,7 @@ void Encoder::resetPosition() {
     _lastPosition = 0;
 }
 
-void Encoder::setResolution(uint16_t resolution) {
+void Encoder::setResolution(uint32_t resolution) {
     _resolution = resolution;
 }
